pyproject.toml
[project]
name = "todolist-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Mohammad",email = "contact.moebra@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)"
]

[tool.poetry]
packages = [{include = "todolist_project", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
